working importer module, needs refactor

+++ src/ucsfbids/modalities/importers/pia/ctpiaimporter.py
@@ -20,10 +20,13 @@ from ucsfbids.modalities.importers.base import CTImporter
 
 
 def strip_json(old_path, new_path):
+    if not old_path.exists():
+        print(f"could not find {old_path}")
+        return
     with open(old_path, "r") as f:
         data_orig = load(f)
 
-    data_clean = {key: value for key, value in data_orig.values() if key not in TO_STRIP}
+    data_clean = {key: value for key, value in data_orig.items() if key not in TO_STRIP}
 
     with open(new_path, "w") as f:
         dump(data_clean, f)
@@ -32,7 +35,9 @@ def strip_json(old_path, new_path):
 TO_STRIP = ["InstitutionName", "InstitutionalDepartmentName", "InstitutionAddress", "DeviceSerialNumber"]
 DEFAULT_FILES = [
     FileSpec("CT", ".nii", Path("CT/CT.nii")),
+    FileSpec("CT", ".nii.gz", Path("CT/CT.nii.gz")),
     FileSpec("CT", ".json", Path("CT/CT_orig.json"), copy_command=strip_json),
+    FileSpec("CT", ".json", Path("CT/CT.json"), copy_command=strip_json),
 ]
 
 
@@ -44,15 +49,44 @@ class CTPiaImporter(CTImporter):
         files: list[FileSpec] = [],
         **kwargs: Any,
     ) -> None:
-        if modality is None:
+        if modality is not None:
             self.modality = modality
 
-        if src_root is None:
+        if src_root is not None:
             self.src_root = src_root
 
         files.extend(DEFAULT_FILES)
         self.files = files
         super().construct(**kwargs)
+
+    def import_all_files(self, path: Path) -> None:
+        assert self.modality is not None
+        assert self.modality.subject_name is not None
+        assert self.src_root is not None
+
+        for file in self.files:
+            subject_name = self.modality.subject_name
+            imaging_root = self.src_root / "data_store2/imaging/subjects"
+            imaging_path = imaging_root / subject_name / file.path_from_root
+            new_path = path / f"{self.modality.full_name}_{file.suffix}{file.extension}"
+            old_name = imaging_path.name
+            exclude = any(n in old_name for n in self.import_exclude_names)
+
+            if new_path.exists():
+                continue
+
+            if exclude:
+                continue
+
+            if imaging_path.is_file():
+                self._import_file(file, imaging_path, new_path)
+                continue
+
+            if not callable(file.copy_command):
+                print(f"No source file for {imaging_path}, skipping...")
+                continue
+
+            file.copy_command(imaging_path, new_path)
 
 
 CT.default_importers["Pia"] = CTPiaImporter
